requirements.txt
django
psycopg2-binary
gunicorn
openai
djangorestframework
djangorestframework-simplejwt
